separated enum tests
